consumer/requirements.txt
requests==2.25.1
kafka-python==2.0.2
psycopg2-binary==2.8.6
aiven-client==2.9.0
pytest==6.2.1
pytest-mock==3.4.0
pytest-cov==2.10.1